requirements.txt
pytorch
nibabel
numpy
pandas
scikit-learn
tensorboard
matplotlib

